initial commit for simple WPF ToDo app.
The application is created in WPF for training purposes.
grid is used for design layout
implemented simple login functionality by name and year.
implemented a simple functionality of adding ToDo tasks to the list of tasks with the output of the date of adding.
minimizing the application to tray is implemented.
implemented additional scripts for creating a watchdog process and saving the application to autoload.

Translated with www.DeepL.com/Translator (free version)